Use a common class for the "specific MOR ETBs with counter" effect
