+++ Mage.Sets/src/mage/cards/g/GuildmagesForum.java
@@ -1,26 +1,22 @@
 package mage.cards.g;
 
-import java.util.UUID;
 import mage.MageObjectReference;
 import mage.abilities.Ability;
 import mage.abilities.costs.common.TapSourceCost;
 import mage.abilities.costs.mana.GenericManaCost;
-import mage.abilities.effects.ReplacementEffectImpl;
+import mage.abilities.effects.common.counter.AddCounterEnteringMOR;
 import mage.abilities.mana.AnyColorManaAbility;
 import mage.abilities.mana.ColorlessManaAbility;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.CardType;
-import mage.constants.Duration;
-import mage.constants.Outcome;
 import mage.constants.WatcherScope;
-import mage.counters.CounterType;
 import mage.game.Game;
-import mage.game.events.EntersTheBattlefieldEvent;
 import mage.game.events.GameEvent;
-import mage.game.permanent.Permanent;
 import mage.game.stack.Spell;
 import mage.watchers.Watcher;
+
+import java.util.UUID;
 
 /**
  *
@@ -66,51 +62,9 @@ class GuildmagesForumWatcher extends Watcher {
                     && event.getSourceId().equals(this.getSourceId())
                     && target != null && target.isCreature(game) && target.getColor(game).isMulticolored()
                     && event.getFlag()) {
-                    game.getState().addEffect(new GuildmagesForumEntersBattlefieldEffect(new MageObjectReference(target.getCard(), game)),
-                        target.getSpellAbility()); //null will cause an immediate crash
+                    game.getState().addEffect(new AddCounterEnteringMOR(new MageObjectReference(target.getCard(), game)),
+                        target.getSpellAbility());
             }
         }
-    }
-
-}
-
-class GuildmagesForumEntersBattlefieldEffect extends ReplacementEffectImpl {
-
-    private final MageObjectReference mor;
-
-    public GuildmagesForumEntersBattlefieldEffect(MageObjectReference mor) {
-        super(Duration.EndOfTurn, Outcome.BoostCreature);
-        this.staticText = "If that mana is spent on a multicolored creature spell, that creature enters the battlefield with an additional +1/+1 counter on it";
-        this.mor = mor;
-    }
-
-    private GuildmagesForumEntersBattlefieldEffect(final GuildmagesForumEntersBattlefieldEffect effect) {
-        super(effect);
-        this.mor = effect.mor;
-    }
-
-    @Override
-    public boolean checksEventType(GameEvent event, Game game) {
-        return event.getType() == GameEvent.EventType.ENTERS_THE_BATTLEFIELD;
-    }
-
-    @Override
-    public boolean applies(GameEvent event, Ability source, Game game) {
-        Permanent permanent = ((EntersTheBattlefieldEvent) event).getTarget();
-        return permanent != null && mor.refersTo(permanent, game);
-    }
-
-    @Override
-    public boolean replaceEvent(GameEvent event, Ability source, Game game) {
-        Permanent target = ((EntersTheBattlefieldEvent) event).getTarget();
-        if (target != null) {
-            target.addCounters(CounterType.P1P1.createInstance(), source.getControllerId(), source, game, event.getAppliedEffects());
-        }
-        return false;
-    }
-
-    @Override
-    public GuildmagesForumEntersBattlefieldEffect copy() {
-        return new GuildmagesForumEntersBattlefieldEffect(this);
     }
 }
